Got fundamental x86 asm compilation to work
