+++ include/io.h
@@ -1,6 +1,10 @@
 #ifndef IO_H
 #define IO_H
 
+#include <stdlib.h>
+#include <stdio.h>
+
+
 char* io_load_file(char* filename);
 char* get_basename(char* path);
 char* get_tempname(char* filename);
@@ -11,5 +15,15 @@ char* get_tempname(char* filename);
         fprintf((fp), (format), ##__VA_ARGS__); \
     } while (0)
 
+#define EXEC(cmd, ...) \
+    do { \
+        char command[256]; \
+        snprintf(command, sizeof(command), (cmd), ##__VA_ARGS__); \
+        printf("Executing: %s\n", command); \
+        int ret = system(command); \
+        if (ret != 0) { \
+            fprintf(stderr, "Error executing command: %s\n", command); \
+        } \
+    } while (0)
 
 #endif // IO_H
