+++ include/x86/instr.h
@@ -9,8 +9,8 @@
 
 void X86_instr(AST_Instruction statement, FILE* fp, X86Stack* stack, bool mep);
 
-void X86_return(AST_Instruction statement, FILE* fp, X86Stack* stack);
-void X86_alloca(AST_Instruction statement, FILE* fp, X86Stack* stack);
+void X86_return(AST_Instruction statement, FILE* fp, X86Stack* stack, bool mep);
+void X86_alloca(AST_Instruction* statement, FILE* fp, X86Stack* stack);
 void X86_assgn(AST_Instruction statement, FILE* fp, X86Stack* stack);
 void x86_binop(AST_Instruction statement, FILE* fp, X86Stack* stack, char** reg);
 void x86_store(AST_Instruction statement, FILE* fp, X86Stack* stack);
